# backend/src/main/resources/openapi.yml
openapi: 3.0.3

info:
  title: UberDup API
  version: 1.0.0

tags:
  - name: User
    description: Operations related to user management
  - name: WalkBud
    description: Operations related to walking bud

paths:
  /user:
    post:
      tags:
        - User
      operationId: createUser
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          description: Bad Request
  /user/{userId}:
    get:
      tags:
        - User
      operationId: getUserData
      summary: Get user information by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
            description: User Info
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/UserData'
        '404':
            description: User Not Found
  /user/deleteUser/{userId}:
    delete:
      tags:
        - User
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User Deleted
        '404':
          description: User Not Found

  /walkingBud/{userId}:
    post:
      tags:
        - WalkBud
      operationId: createWalkingBud
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Walking Bud Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkingBudCreatedResponse'
        '404':
          description: User Not Found to create

  /walkingBud/{walkerId}:
    post:
      tags:
        - WalkBud
      operationId: deleteWalkingBud
      parameters:
        - name: walkerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Walker deleted successfully
        '400':
          description: Walker not found

components:
  schemas:
    WalkingBudCreatedResponse:
      type: object
      properties:
        walkerId:
          type: string
        name:
          type: string

    Sex:
      type: string
      enum:
        - M
        - F
        - OTHER

    UserCreateRequest:
      type: object
      required:
        - email
        - name
        - password
        - residence
        - sex
        - age
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
          format: password
          minLength: 8
          writeOnly: true
        residence:
          $ref: '#/components/schemas/Habitation'
        sex:
          $ref: '#/components/schemas/Sex'
        age:
          type: integer
          format: int32
          minimum: 0

    UserCreatedResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string

    UserData:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        residence:
          $ref: '#/components/schemas/Habitation'
        sex:
          $ref: '#/components/schemas/Sex'
        age:
          type: integer
          format: int32
          minimum: 0

    Habitation:
      type: object
      required:
        - street
        - number
        - city
        - state
        - country
        - zipCode
      properties:
        street:
          type: string
        number:
          type: string
        apartment:
          type: string
        reference:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zipCode:
          type: string